git-svn-id: svn://123.57.173.36/jhj-parent2.0/Branches/jhj-parent2.1@326 57d5540f-7636-4a73-8af3-7dc9f519f443

+++ jhj-service/src/main/java/com/jhj/service/bs/OrgStaffFinanceService.java
@@ -16,9 +16,9 @@ import com.jhj.vo.OrgStaffFinanceSearchVo;
 public interface OrgStaffFinanceService {
 	int deleteByPrimaryKey(Long id);
 
-    int insert(OrgStaffFinance record);
+    Long insert(OrgStaffFinance record);
 
-    int insertSelective(OrgStaffFinance record);
+    Long insertSelective(OrgStaffFinance record);
 
     OrgStaffFinance selectByPrimaryKey(Long id);
